# Merge elementwise implementations