--- a/cupy/core/_reduction.pxd
+++ b/cupy/core/_reduction.pxd
@@ -27,7 +27,7 @@ cdef class _AbstractReductionKernel:
 
     cdef function.Function _get_function(
         self,
-        tuple params, tuple arginfos, _kernel._TypeMap types,
+        tuple params, tuple args_info, _kernel._TypeMap type_map,
         str map_expr, str reduce_expr, str post_map_expr, str reduce_type,
         Py_ssize_t block_size)