--- a/cupy/core/_reduction.pxd
+++ b/cupy/core/_reduction.pxd
@@ -29,7 +29,7 @@ cdef class _AbstractReductionKernel:
         self,
         tuple params, tuple args_info, _kernel._TypeMap type_map,
         str map_expr, str reduce_expr, str post_map_expr, str reduce_type,
-        Py_ssize_t block_size)
+        Py_ssize_t block_size, int device_id)
 
 
 cdef class ReductionKernel(_AbstractReductionKernel):