--- a/cupy/core/_reduction.pyx
+++ b/cupy/core/_reduction.pyx
@@ -5,7 +5,6 @@ from cupy.core._dtype cimport get_dtype
 from cupy.core cimport _kernel
 from cupy.core._kernel cimport _broadcast
 from cupy.core._kernel cimport _check_array_device_id
-from cupy.core._kernel cimport _get_arginfos
 from cupy.core._kernel cimport _get_kernel_params
 from cupy.core._kernel cimport _get_out_args
 from cupy.core._kernel cimport _get_out_args_with_params
@@ -21,6 +20,9 @@ from cupy.core.core cimport ndarray
 from cupy.core cimport internal
 from cupy.cuda cimport device
 from cupy.cuda cimport function
+from cupy.cuda.function cimport Arg
+from cupy.cuda.function cimport ScalarArg
+from cupy.cuda.function cimport IndexerArg
 from cupy.cuda cimport runtime
 
 import string
@@ -34,7 +36,7 @@ from cupy import util
 
 
 cpdef function.Function _create_reduction_function(
-        name, block_size, reduce_type, params, arginfos, identity,
+        name, block_size, reduce_type, params, args, identity,
         pre_map_expr, reduce_expr, post_map_expr,
         _kernel._TypeMap type_map, input_expr, output_expr, preamble, options):
     module_code = string.Template('''
@@ -94,7 +96,7 @@ extern "C" __global__ void ${name}(${params}) {
         name=name,
         block_size=block_size,
         reduce_type=reduce_type,
-        params=_kernel._get_kernel_params(params, arginfos),
+        params=_kernel._get_kernel_params(params, args),
         identity=identity,
         reduce_expr=reduce_expr,
         pre_map_expr=pre_map_expr,
@@ -234,6 +236,10 @@ cdef class _AbstractReductionKernel:
         cdef Py_ssize_t block_size, block_stride, out_block_num
         cdef ndarray ret
         cdef function.Function kern
+        cdef Arg arg
+
+        # TODO(niboshi): Replace in_args with this
+        in_args_ = [Arg.from_obj(a) for a in in_args]
 
         if dtype is not None:
             dtype = get_dtype(dtype).type
@@ -258,7 +264,7 @@ cdef class _AbstractReductionKernel:
 
         out_shape = _get_out_shape(a_shape, reduce_axis, out_axis, keepdims)
         out_args = self._get_out_args(out_args, out_types, out_shape)
-        ret = out_args[0]
+        ret = (<Arg>out_args[0]).obj
         if ret.size == 0:
             return ret
 
@@ -266,9 +272,10 @@ cdef class _AbstractReductionKernel:
             raise ValueError(('zero-size array to reduction operation'
                               ' %s which has no identity') % self.name)
 
-        in_args = [x if isinstance(x, ndarray) else
-                   _scalar.CScalar.from_numpy_scalar_with_dtype(x, t)
-                   for x, t in zip(in_args, in_types)]
+        for i in range(self.nin):
+            arg = in_args_[0]
+            if arg.is_scalar():
+                (<ScalarArg>arg).apply_dtype(in_types[i])
         in_shape = _set_permuted_args(
             in_args, reduce_axis + out_axis, a_shape, self.in_params)
 
@@ -285,20 +292,21 @@ cdef class _AbstractReductionKernel:
             contiguous_size)
 
         # Kernel arguments passed to the __global__ function.
-        inout_args = (
-            in_args
+        inout_args = tuple(
+            in_args_
             + out_args
             + [
-                _carray.Indexer(in_shape),
-                _carray.Indexer(out_shape),
+                IndexerArg(in_shape),
+                IndexerArg(out_shape),
                 # block_stride is passed as the last argument.
-                _scalar.CScalar.from_int32(block_stride),
+                #_scalar.CScalar.from_int32(block_stride),
+                Arg.from_obj(numpy.int32(block_stride)),
             ])
 
         # Retrieve the kernel function
         func = self._get_function(
             self._params,
-            _get_arginfos(inout_args),
+            inout_args,
             type_map,
             map_expr, reduce_expr, post_map_expr, reduce_type,
             block_size)
@@ -319,7 +327,7 @@ cdef class _AbstractReductionKernel:
 
     cdef function.Function _get_function(
             self,
-            tuple params, tuple arginfos, _kernel._TypeMap type_map,
+            tuple params, tuple args, _kernel._TypeMap type_map,
             str map_expr, str reduce_expr, str post_map_expr, str reduce_type,
             Py_ssize_t block_size):
         raise NotImplementedError()
@@ -374,7 +382,6 @@ cdef class _SimpleReductionKernel(_AbstractReductionKernel):
             raise TypeError(
                 'Argument \'a\' has incorrect type (expected %s, got %s)' %
                 (ndarray, type(a)))
-        in_args = [arr]
 
         dev_id = device.get_device_id()
         _check_array_device_id(arr, dev_id)
@@ -385,17 +392,21 @@ cdef class _SimpleReductionKernel(_AbstractReductionKernel):
             _check_array_device_id(out, dev_id)
             out_args = [out]
 
-        reduce_dims = True
         return self._call(
-            in_args, out_args,
-            arr.shape, axis, dtype, keepdims, reduce_dims, None)
+            [arr],  # in_args
+            out_args,
+            arr.shape, axis, dtype, keepdims,
+            True,  # reduce_dims
+            None)
 
     cdef tuple _get_expressions_and_types(
             self, list in_args, list out_args, dtype):
         cdef _kernel._Op op
 
+        # TODO(niboshi): Do this earlier
+        in_args_ = [Arg.from_obj(a) for a in in_args]
         op = self._ops.guess_routine(
-            self.name, self._routine_cache, in_args, dtype, self._ops)
+            self.name, self._routine_cache, in_args_, dtype, self._ops)
         map_expr, reduce_expr, post_map_expr, reduce_type = op.routine
 
         if reduce_type is None:
@@ -423,27 +434,42 @@ cdef class _SimpleReductionKernel(_AbstractReductionKernel):
 
     cdef function.Function _get_function(
             self,
-            tuple params, tuple arginfos, _kernel._TypeMap type_map,
+            tuple params, tuple args, _kernel._TypeMap type_map,
             str map_expr, str reduce_expr, str post_map_expr, str reduce_type,
             Py_ssize_t block_size):
         return _SimpleReductionKernel_get_cached_function(
             map_expr, reduce_expr, post_map_expr, reduce_type,
-            params, arginfos, type_map,
+            params, args, type_map,
             self.name, block_size, self.identity,
             self._input_expr, self._output_expr, self._preamble, ())
 
 
-@util.memoize(for_each_device=True)
+_simple_reduction_kernel_func_cache = {}
+
 def _SimpleReductionKernel_get_cached_function(
         map_expr, reduce_expr, post_map_expr, reduce_type,
-        params, arginfos, _kernel._TypeMap type_map,
+        params, args, _kernel._TypeMap type_map,
         name, block_size, identity, input_expr, output_expr, _preamble,
         options):
+    key = (
+        map_expr, reduce_expr, post_map_expr, reduce_type,
+        params,
+        tuple([(<Arg>arg).get_immutable_key() for arg in args]),
+        type_map,
+        name, block_size, identity, input_expr, output_expr, _preamble,
+        options
+    )
+    func = _simple_reduction_kernel_func_cache.get(key)
+    if func is not None:
+        return func
 
-    return _create_reduction_function(
-        name, block_size, reduce_type, params, arginfos, identity,
+    func = _create_reduction_function(
+        name, block_size, reduce_type, params, args, identity,
         map_expr, reduce_expr, post_map_expr,
         type_map, input_expr, output_expr, _preamble, options)
+
+    _simple_reduction_kernel_func_cache[key] = func
+    return func
 
 
 # -----------------------------------------------------------------------------
@@ -553,8 +579,8 @@ cdef class ReductionKernel(_AbstractReductionKernel):
             out_args = [out]
 
         dev_id = device.get_device_id()
-        in_args = _preprocess_args(dev_id, args[:self.nin], False)
-        out_args = _preprocess_args(dev_id, out_args, False)
+        in_args = _preprocess_args(dev_id, args[:self.nin])
+        out_args = _preprocess_args(dev_id, out_args)
         in_args, broad_shape = _broadcast(in_args, self.in_params, False)
 
         return self._call(
@@ -586,29 +612,40 @@ cdef class ReductionKernel(_AbstractReductionKernel):
 
     cdef function.Function _get_function(
             self,
-            tuple params, tuple arginfos, _kernel._TypeMap type_map,
+            tuple params, tuple args, _kernel._TypeMap type_map,
             str map_expr, str reduce_expr, str post_map_expr, str reduce_type,
             Py_ssize_t block_size):
         return _ReductionKernel_get_cached_function(
-            self.nin, self.nout, params, arginfos, type_map,
+            self.nin, self.nout, params, args, type_map,
             self.name, block_size, reduce_type, self.identity,
             map_expr, reduce_expr, post_map_expr,
             self.preamble, self.options)
 
 
-@util.memoize(for_each_device=True)
+_reduction_kernel_func_cache = {}
+
 def _ReductionKernel_get_cached_function(
-        nin, nout, params, arginfos, _kernel._TypeMap type_map,
+        nin, nout, params, args, _kernel._TypeMap type_map,
         name, block_size, reduce_type, identity, map_expr, reduce_expr,
         post_map_expr, preamble, options):
+    key = (
+        device.get_device_id(),
+        nin, nout, params,
+        tuple([(<Arg>arg).get_immutable_key() for arg in args]),
+        type_map, name, block_size, reduce_type, identity,
+        map_expr, reduce_expr, post_map_expr, preamble, options)
+    func = _reduction_kernel_func_cache.get(key)
+    if func is not None:
+        return func
+
     cdef _kernel.ParameterInfo p
-    cdef _kernel._ArgInfo arginfo
+    cdef function.Arg arg
     in_arrays = [
-        p for p, arginfo in zip(params[:nin], arginfos[:nin])
-        if not p.raw and arginfo.is_ndarray()]
+        p for p, arg in zip(params[:nin], args[:nin])
+        if not p.raw and arg.is_ndarray()]
     out_arrays = [
-        p for p, arginfo in zip(params[nin:nin+nout], arginfos[nin:nin+nout])
-        if not p.raw and arginfo.is_ndarray()]
+        p for p, arg in zip(params[nin:nin+nout], args[nin:nin+nout])
+        if not p.raw and arg.is_ndarray()]
     input_expr = '\n'.join(
         [(('const {0} {1}' if p.is_const else '{0}& {1}') +
           ' = _raw_{1}[_in_ind.get()];').format(p.ctype, p.name)
@@ -617,7 +654,10 @@ def _ReductionKernel_get_cached_function(
         ['{0} &{1} = _raw_{1}[_out_ind.get()];'.format(p.ctype, p.name)
          for p in out_arrays if not p.is_const])
 
-    return _create_reduction_function(
-        name, block_size, reduce_type, params, arginfos, identity,
+    func = _create_reduction_function(
+        name, block_size, reduce_type, params, args, identity,
         map_expr, reduce_expr, post_map_expr,
         type_map, input_expr, output_expr, preamble, options)
+
+    _reduction_kernel_func_cache[key] = func
+    return func