… cupy/cuda/function.pxd
… cupy/cuda/function.pyx
… cupy/testing/helper.py