Elementwise version of #2732.
Merges the logic of `ElementwiseKernel` and `ufunc`.