@niboshi 2020-01-09
@emcastillo 
I think refactoring on elementwise-kernels is finished. May I hear a brief comment on the last commit (b746a0e791d765fabe5b7ddf36c14e8a1164a040)?
Some code is still duplicated (e.g. `_preprocess_args`) because it's also used from `reduction.pyx`.
If basic idea is OK, I'll try to update `reduction.pyx` too.
Reconsideration may happen at that moment.