@emcastillo 2020-01-09
At first glance, the PR changes look fine and consistent.
 But one of the things I am worried about is that with `ufunc` and `elementwise` being merged, it may complicate adding support for [generalized functions](https://docs.scipy.org/doc/numpy-1.13.0/reference/c-api.generalized-ufuncs.html) since these can do reductions as well.

In numpy ufuncs are a special case of the generalized ufunc. (described by just the signature).
So I think that having a separated interface for the ufunc and argument checkings, signatures etc decoupled from the Elementwise kernel may keep things simpler. However, my knowledge of all these internals is limited and I am not sure of to what extent they can be a problem. @toslunar @asi1024 any comments on this?.