@niboshi 2020-01-09
@emcastillo 
I agree that element kernels and reduction kernels can be unified.
However, as it's difficult to do that all at once, you can consider this PR as a first step toward that.
(currently I don't have a concrete idea how they all will be unified though.)

>So I think that having a separated interface for the ufunc and argument checkings, signatures etc decoupled from the Elementwise kernel may keep things simpler.

I don't understand this point.
My hope is all the kernel implementations (`ufunc`, `ElementwiseKernel`, `_SimpleReductionKernel`, `ReductionKernel`) should be able to be generalized to some extent.
Could you elaborate why you think `ufunc` and `ElementwiseKernel` should be separate?