@emcastillo 2020-01-09
If reduction and elementwise are going to be unified then there is no problem at all 😄