@emcastillo 2020-01-09
This is a long one, so I just want to be sure that I am getting everything right. 
What you have done here is:
1) Rewrite the kernel argument managing logic and abstract it into several classes instead of raw tuples and module level functions.
2) Create a base _AbstractElementwiseKernel that essentially groups all the ufunc related `_kernel` module functions.

I think it might be better to split this in two PRs, one with just `_ArgKind`, `_ArgInfo` and `_TypeMap` and another one with the rest of changes. But if you prefer to keep them as one PR I won't be against it.