@niboshi 2020-01-09
@emcastillo 
Thank you for a comment.
I think your suggestion is very right. I'll separate 1. as another PR while this PR is kept (I mean this PR will depend on it).