@niboshi 2020-01-09
Btw, `_Args` class will be classified in 2., as it's abstraction between `ufunc` and `ElementwiseKernel`.