@niboshi 2020-02-27
After some internal discussion, we have decided to include scalar refactoring (#2918) in this PR, because it's difficult to discuss separately.

I'm planning to introduce `Arg` class to wrap arguments and provide uniform interface across types/kernel implementations. `Arg.get_pointer()` which returns `CPointer` is one example.

My current plan (although subject to change) is:

* Implement `Arg` in `function.pyx`. Note that `Function.__call__` has to accept `Arg`s as well as raw ndarrays and scalars.
* Merge `_ArgInfo` into `Arg`. Note that `_ArgInfo` was meant to be hashable but `Arg` is not. So, I am considering adding something like `Arg.get_immutable_key()`.