@niboshi 2020-02-28
Reduction is still wip, but elementwise tests pass.