@asi1024 2020-03-25
(Discussed offline with @okuta)
This PR conflicts with #2702 and merged it first, but performance improvement of this kind should be applied after this PR is merged. So, I decided to revert #2702 in this PR temporarily and send the performance improvement PR later.