@emcastillo 2020-03-27
Benchmarked ufuncs,

Consistent increase in CPU time from 10us to 20us
This PR
```
time_sqrt            - case        100                    :    CPU:   20.238 us   +/- 0.263 (min:   19.956 / max:   20.900) us     GPU:   24.029 us   +/- 0.731 (min:   23.264 / max:   25.504) us
```
Master
```
time_sqrt            - case        100                    :    CPU:   11.371 us   +/- 0.247 (min:   11.143 / max:   11.853) us     GPU:   14.662 us   +/- 1.109 (min:   13.984 / max:   17.888) us
```

This time diff holds for all ufuncs, elementwise kernels, reduction and raw kernels.