@kmaehashi 2020-04-23
Reverted in #3296 for performance issue.